requirements.txt
torch
transformers
torchvision
Pillow
numpy
matplotlib
pandas